Complete JWT authentication and create post feature

+++ src/main/java/com/example/blogkar/post/dto/PostResponse.java
@@ -12,7 +12,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 public class PostResponse {
 
-    private long postId;
+    private Integer postId;
 
     private String title;
