+++ src/main/java/com/example/blogkar/post/repository/PostRepository.java
@@ -3,6 +3,7 @@ package com.example.blogkar.post.repository;
 import com.example.blogkar.post.entity.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostRepository extends JpaRepository<Post, Long> {
+public interface PostRepository extends JpaRepository<Post, Integer
+        > {
     boolean existsBySlug(String slug);
 }
